refactor: 프로젝트 구조 수정 및 상수 활용
- 공통적으로 사용되는 값들은 상수로 관리
- 전부 baseball 디렉토리 내부에서 관리
