+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import domain.Game;
+import baseball.domain.Game;
 
 public class Application {
     public static void main(String[] args) {
