+++ src/main/java/baseball/Constants.java
@@ -1,0 +1,11 @@
+package baseball;
+
+public class Constants {
+    public final static String OUT_OF_VALID_NUM_RANGE_ERROR = "잘못된 숫자를 입력하셨습니다.";
+    public final static String NOT_NUMBER_TYPE_ERROR = "잘못된 값을 입력하셨습니다.";
+    public final static int ANSWER_LIST_LENGTH = 3;
+    public final static int RANGE_START_NUM = 1;
+    public final static int RANGE_END_NUM = 9;
+    public final static int RESTART_NUM = 1;
+    public final static int END_GAME_NUM = 2;
+}
